fix: Line status now namespaced
